Update templates for use with vSphere builder.
This removes the post-processors and adds new environment variables for:

- `ESX_HOST`
- `ESX_USERNAME`
- `ESX_PASSWORD`
- `DATASTORE`

Additionally, this adds packer variables for `vm_username` and
`vm_password` which default to `vagrant` and can be changed by setting
the value of `PACKER_VARS` when running the make command.

It should be noted that the post-processors are not needed when
building directly on an ESX host, and instating a one will result in
a failing packer run as packer attempts to process the files on the
local system rather than the ESX host itself.