Make buildkite script branch and commit id aware.
This now checks to determine changes between master and a given commit hash when no passed builds for a given branch exist.

Additionally, we now check the local repo to determine if the hash of the last passed build still exists in the event of a force-push. This will cause the script to behave as if it is a new branch, and compare against master as it likely will not have a valid commit ID to compare against.