Generate a new UUID for each make invocation.
Packer insists on havinging a uniqe output directory for each generated
system. While using the BUILDKITE_BUILD_ID value would make it helpful
to identify objects to a given build, it causes collitions at the
moment. This forces a new UUID to be generated for each invocation, and
assumes the make command is run in serial to result in the correct
output. We will need to optimize this later.