Build systems serially.
This is less than ideal as it is very slow. However, we need to get
this working and optimize later.

Exit with non-zero status if a build fails.

The Ruby script would always exit with a zero status and incorrectly
indicate a successful build even when it was not. This fixes the issue.