Add WMF upgrade script and remove patches from 2k8
This adds a script for installing WMF (PowerShell 4.x) and .net 4.5.
The script which I originally copied did not download the required
installer or update file . This downloads the correct file for the
architecture and also ensures we do not restart to allow packer to
complete the entire process. Additionally, a simple powershell
script to download files was added as wget.exe which is downloaded
does not have a valid SSL certificate chain nor access to the
one built-in to Windows, and as such has no way of validating the
URL with this tool, whereas PowerShell can use the Windows built-in
and do so.